Extract _write_layout.py from _writer.py (#2248)
Move TIFF/BigTIFF IFD encoding primitives and the BigTIFF / COG layout
planners out of _writer.py into a new _write_layout.py. The pixel
encoding kernels (strip/tile compression, photometric/predictor encode)
and the top-level write entry points stay in _writer.py.

Moved: _float_to_rational, _serialize_tag_value, _pack_tag_value,
_build_ifd, _assemble_tiff, _promote_offsets_to_long8,
_assemble_standard_layout, _assemble_cog_layout,
_compute_classic_ifd_overhead, _should_use_bigtiff_streaming, the BO
byte-order constant, and the _BIGTIFF_OFFSET_TAGS frozenset.

_writer.py re-exports every moved name so call sites in
_writers/eager.py, _writers/gpu.py, _gpu_decode.py, and the test suite
keep working. _assemble_tiff resolves the layout helpers and
_resolve_photometric / _OVERRIDABLE_AUTO_TAG_IDS /
_DANGEROUS_EXTRA_TAG_IDS through the _writer module at call time so
monkeypatches on _writer.* (e.g. test_eager_bigtiff_overhead_exact_1905)
still take effect.

_writer.py drops about 580 lines. Behavior-neutral; the geotiff test
suite passes.

Part of #2211.
